+++ doc/whats-new.md
@@ -21,6 +21,8 @@ What's new
 
 ### New features
 
+- Option to reverse sign of toroidal magnetic field Bt for TokamakEquilibrium (#114)\
+  By [John Omotani](https://github.com/johnomotani)
 - Enable DCT interpolation scheme in TokamakEquilibrum, can be selected with new option
   psi_interpolation_method (#113)\
   By [John Omotani](https://github.com/johnomotani)